+++ lib/create.js
@@ -3,6 +3,8 @@ const path = require('path');
 const inquirer = require('inquirer');
 const chalk = require('chalk');
 
+const files = require('./files');
+
 const TemplatingService = require('../service/templating.service');
 
 const TEMPLATE_CHOICES = fs.readdirSync(path.join(__dirname, '../templates'));
@@ -36,32 +38,33 @@ const generateProjectName = () => {
 
 const getProjectOptions = async (project) => {
   const answers = await inquirer.prompt(QUESTIONS)
-  console.log(answers);
 
-  const projectChoice = answers['template'];
+  const templateName = answers['template'];
   const githubUsername = answers['github'];
   const projectName = project ? project : generateProjectName();
-
-  const CURR_DIR = process.cwd();
-  const templatePath = path.join(__dirname, '../templates', projectChoice);
-  const targetPath = path.join(CURR_DIR, projectName);
 
   const options = {
     projectName,
     githubUsername,
-    templateName: projectChoice,
-    templatePath,
-    targetPath,
+    templateName
   }
   return options
 }
 
+const checkForProjectPathConflict = (project) => {
+  // returns false if no conflict
+  return files.checkFileOrDirExists(path.join(process.cwd(), project));
+}
+
 module.exports = {
   newProject: async (project) => {
+    const conflict = checkForProjectPathConflict(project)
+    if (conflict) {
+      console.log(chalk.red(`Whoops! Project with name '${project}' already exists.\nTry again with a different project name.`))
+      return false
+    }
     const options = await getProjectOptions(project);
-    console.log(options)
-    const { templateName } = options;
     const templatingService = new TemplatingService();
-    templatingService.template(templateName, { projectName: project });
+    templatingService.template({...options});
   },
 }
