+++ service/templating.service.js
@@ -3,51 +3,19 @@ const path = require('path');
 const tmp = require('tmp');
 const ejs = require('ejs');
 
-/**
- * @typedef {Object} TemplateOptions
- * @property {string} projectName
- */
+const walk = require('../lib/files').walk;
+const ensureDirectoryExistence = require('../lib/files').ensureDirectoryExistence
 
-// taken from https://stackoverflow.com/a/16684530/1204556
-const walk = function(dir) {
-  let results = [];
-  const list = fs.readdirSync(dir);
-  list.forEach(function(pathFromDirectoryRoot) {
-      const absolutePath = dir + '/' + pathFromDirectoryRoot;
-
-      const stat = fs.statSync(absolutePath);
-      if (stat && stat.isDirectory()) {
-          /* Recurse into a subdirectory */
-          results = results.concat(walk(absolutePath));
-      } else {
-          /* Is a file */
-          results.push({
-            absolutePath
-          });
-      }
-  });
-
-  return results;
-}
-
-const setRelativePaths = (dir, files=[]) => files.map(({ absolutePath }) => ({
+const setRelativePaths = (dir, files=[]) => (
+  files.map(({ absolutePath }) => ({
   absolutePath,
   pathFromDirectoryRoot: path.relative(dir, absolutePath)
-}));
+  }))
+);
 
-const getTemplatesPath = (name) => path.join(__dirname, '..', 'templates', name);
+const getTemplatePath = (name) => path.join(__dirname, '../templates', name);
 const getTemporaryDirectory = () => tmp.dirSync();
-
-const getTarget = projectName => path.join(process.cwd(), projectName);
-
-const checkFileOrDirExists = path => fs.existsSync(path);
-
-function ensureDirectoryExistence(filePath) {
-  var dirname = path.dirname(filePath);
-  if (!fs.existsSync(dirname)) {
-    fs.mkdirSync(dirname, { recursive: true });
-  }
-}
+const getTargetPath = projectName => path.join(process.cwd(), projectName);
 
 /**
  * @param {TemplateOptions} options
@@ -61,10 +29,8 @@ const validateOptions = options => {
     throw new Error("No project name specified");
   }
 
-  const target = getTarget(options.projectName);
-  const conflict = checkFileOrDirExists(target);
-  if (conflict) {
-    throw new Error("Whoops! A file or directory by the name already exists at " + projectName);
+  if (!options.templateName) {
+    throw new Error("No template name specified");
   }
 }
 
@@ -74,11 +40,11 @@ module.exports = class TemplatingService {
    * @param {*} templateName
    * @param { TemplateOptions } options
    */
-  template(templateName = 'vanilla' , options = {}) {
-    console.log({templateName, options})
-    // validateOptions(options);
+  template(options = {}) {
+    validateOptions(options);
+    const { templateName } = options;
 
-    const templatesPath = getTemplatesPath(templateName);
+    const templatesPath = getTemplatePath(templateName);
     const files = setRelativePaths(templatesPath, walk(templatesPath));
     const temporaryDirectory = getTemporaryDirectory().name;
 
@@ -90,7 +56,7 @@ module.exports = class TemplatingService {
       fs.writeFileSync(tmpDestination, output);
     });
 
-    const finalTarget = getTarget(options.projectName);
+    const finalTarget = getTargetPath(options.projectName);
     ensureDirectoryExistence(finalTarget);
     fs.copySync(temporaryDirectory, finalTarget);
   }
