--- test-proj/index.html
@@ -1,19 +1,0 @@
-<!DOCTYPE html>
-<html lang="en">
-  <head>
-    <meta charset="UTF-8" />
-    <meta name="viewport" content="width=device-width, initial-scale=1.0" />
-    <title>test-proj</title>
-    <link rel="stylesheet" href="./styles/style.css" />
-  </head>
-  <body>
-    <main class="body-outer">
-      <div class="body-inner">
-        <h1>test-proj</h1>
-        <p>Counter: <span id="output">0</span></p>
-        <button id="increment-button">Increment</button>
-      </div>
-    </main>
-    <script src="./src/index.js"></script>
-  </body>
-</html>
